+++ scripts/MageMagicAddon.js
@@ -10,9 +10,9 @@ export class MageMagicAddon {
     MTA_MANA: "swnr-space-magic-mta-mana",
     MTA_ARCANA_IMPORTANCE: "arcana-importance",
     ACTOR_SHEET_THEME: "sheet-theme",
-    ACTOR_SPELL_FILTER_ARCANUM: 'spell-filter-arcanum',
-    ACTOR_SCENE_PARADOX: 'scene-paradox',
-    ACTOR_MORALITY: 'morality',
+    ACTOR_SPELL_FILTER_ARCANUM: "spell-filter-arcanum",
+    ACTOR_SCENE_PARADOX: "scene-paradox",
+    ACTOR_MORALITY: "morality",
     //Spells
     MTA_SPELL_ARCANUM: "arcanum",
     MTA_SPELL_PRACTICE: "practice",
@@ -42,6 +42,8 @@ export class MageMagicAddon {
     ROLLCOD_CHAT: `modules/${this.ID}/templates/chat/roll-cod.hbs`,
     ROLLSWNR: `modules/${this.ID}/templates/dialogs/roll-swnr.hbs`,
     ROLLSWNR_CHAT: `modules/${this.ID}/templates/chat/roll-swnr.hbs`,
+    ROLLWISDOM: `modules/${this.ID}/templates/dialogs/roll-wisdom.hbs`,
+    ROLLWISDOM_CHAT: `modules/${this.ID}/templates/chat/roll-wisdom.hbs`,
   };
 
   //   static initialize() {